Fix build
This was copy from https://github.com/mroy/stellaris-launchpad-template-gcc.git
This fix errors that some stdio function are getting because of
leak of _sbrk().

TODO: printf primitives are sending char over uart.
See if this uart is the virtual uart that send
string over virtal port.
If not change to that or remove