Some motors handling
Add a message to arm esc's, this will send a 1000micro high pulse to esc's
to arm they in the lowest configuration, with this value motors will not
spin.

Also add a flag to motors_velocity_set() to filter the values send to motors.
When quad is fling we can not send a 1000 value, this will turn motors off and
quad will fall.

So in fly we have a value that will keep all four motors spinning.