Change PID to receive integers instead of floats
This way we will not have issue with overcorrection and
when lifting off the quad should be more stable.

One of issue find in last test is that quad always start lift in the same side.
I had try balance weight but not work, this change should fix.
As error was arround 0.6 and this generate a lower proportional gain but
the integral after some time turn into a big gain.